Refactor   *add link to css in html files   *create food model and extract food functions from index   *WIP - create food events file and extract
