+++ foods.html
@@ -4,6 +4,7 @@
   <meta charset="utf-8">
   <title>Quantified Self - Manage Foods</title>
   <link rel="stylesheet" href="//maxcdn.bootstrapcdn.com/font-awesome/4.3.0/css/font-awesome.min.css">
+  <link rel="stylesheet" href="./lib/stylesheets/index.css">
 </head>
 <body>
   <img src="public/palm-pattern.png" alt="Red Palm Trees">
