+++ index.html
@@ -5,6 +5,7 @@
   <title>Quantified Self - Diary</title>
   <link href="https://fonts.googleapis.com/css?family=Yantramanav:400,700|Zilla+Slab+Highlight:700" rel="stylesheet">
   <link rel="stylesheet" href="//maxcdn.bootstrapcdn.com/font-awesome/4.3.0/css/font-awesome.min.css">
+  <link rel="stylesheet" href="./lib/stylesheets/index.css">
 </head>
 <body class="index-home">
   <header id="header">
