+++ lib/index.js
@@ -1,77 +1,5 @@
-require('./stylesheets/index')
-var sheets = require('./styles.css');
 const $ = require('jquery')
-//var Food = require('../models/food')
-
-function Food(food) {
-  this.id = food.id;
-  this.name = food.name;
-  this.calories = food.calories;
-}
-
-Food.prototype.toHTML = function() {
-  return `<tr class="food_row">` +
-  `<td class="food-name" data-id="${this.id}">${this.name}</td>`+
-  `<td class="food_calories">${this.calories}</td>`+
-  `<td><a><input type="image" src="public/trash_can.png" class="trash-can" id=${this.id} alt="a trash can"/></a></td></tr>`
-}
-
-Food.prototype.toHTMLDiary = function(data) {
-  return `<tr class="food_row">` +
-  `<td class="food-name" data-id="${this.id}">${this.name}</td>`+
-  `<td class="food_calories">${this.calories}</td>`+
-  `<td><input class="checkbox" type="checkbox"</td></tr>`
-}
-
-Food.getAllFoods = function() {
-  return $.ajax({
-    type: "GET",
-    url: 'http://localhost:3000/api/v1/foods',
-    dataType:"json",
-  })
-  .done(function(data) {
-    return data;
-  })
-  .fail(function(error) {
-    debugger
-  })
-}
-
-Food.allFoodsToHTML = function() {
-  return this.getAllFoods()
-  .then(function(foodList) {
-    return foodList.map(function(food) {
-      return new Food(food);
-    }).sort(function(a, b) {
-      return b.id - a.id;
-    })
-  })
-  .then(function(foodList) {
-    return foodList.map(function(food) {
-      if($('.foods_table').length) {
-        // debugger
-        return food.toHTML();
-      } else {
-        // debugger
-        return food.toHTMLDiary();
-      }
-      // return food.toHTML();
-    })
-  })
-}
-
-Food.delete = function(foodID) {
-  return $.ajax({
-    url: `http://localhost:3000/api/v1/foods/` + foodID,
-    type: 'DELETE',
-    dataType: "json",
-    error: function(error) {
-      // alert("Cannot delete food")
-    }
-  }).done(function(response) {
-    return response
-    })
-}
+const Food = require("./models/food")
 
 $('#new_foods').submit(function(event) {
   event.preventDefault()
@@ -125,16 +53,10 @@ $('.foods_table').on('click', '.trash-can', function(event) {
   Food.delete(foodID)
   Food.removeFood(foodID);
 });
-
-Food.removeFood = function(element) {
-  $('.food_row').each(function() {
-    if (this.firstChild.attributes['data-id'].value == element) {
-      this.remove()
-    }
-  })
-}
+//require('./events/food-events')
 
 $( document ).ready(function() {
+
     Food.allFoodsToHTML()
     .then(function(foodHTML) {
       $('#foods_table').html(foodHTML)
